+++ include/vrv/calcdotsfunctor.h
