+++ include/vrv/calcledgerlinesfunctor.h
