+++ src/findfunctor.cpp
@@ -5,7 +5,7 @@
 // Copyright (c) Authors and others. All rights reserved.
 /////////////////////////////////////////////////////////////////////////////
 
-#include "find.h"
+#include "findfunctor.h"
 
 //----------------------------------------------------------------------------
