+++ src/system.cpp
@@ -19,8 +19,7 @@
 #include "doc.h"
 #include "dynam.h"
 #include "ending.h"
-#include "find.h"
-#include "functor.h"
+#include "findfunctor.h"
 #include "functorparams.h"
 #include "layer.h"
 #include "measure.h"
